refactor: Header 컴포넌트 구조 변경 및 리팩토링
