+++ src/features/header-elements/view-type-toggle-button.tsx
@@ -1,0 +1,24 @@
+import { LayoutGrid, List } from 'lucide-react'
+
+export const ViewTypeToggleButton = ({
+  isListView,
+  onClick,
+}: {
+  isListView: boolean
+  onClick: () => void
+}) =>
+  isListView ? (
+    <LayoutGrid
+      size={17}
+      strokeWidth={1.5}
+      onClick={onClick}
+      className='cursor-pointer'
+    />
+  ) : (
+    <List
+      size={18}
+      strokeWidth={1.5}
+      onClick={onClick}
+      className='cursor-pointer'
+    />
+  )
