+++ src/features/index.ts
@@ -8,3 +8,4 @@ export * from './place-review'
 export * from './course-plan-card'
 export * from './notification'
 export * from './main'
+export * from './header-elements'
