+++ src/shared/ui/index.ts
@@ -8,3 +8,4 @@ export * from './region-cascader'
 export * from './select-categories'
 export * from './spacer'
 export * from './tab-button'
+export * from './back-button'
