+++ src/views/list-place-review/index.tsx
@@ -3,7 +3,7 @@
 import { useGetPlaceReviews } from '@/src/entities/place'
 import ReviewCommentCard from '@/src/widgets/review-comment-card'
 import { Spacer } from '@/src/shared/ui'
-import { HeaderWithBackButton } from '@/src/widgets/header'
+import { ActionHeader } from '@/src/widgets'
 
 export default function ListPlaceReview({ placeId }: { placeId: string }) {
   const { data: reviewData } = useGetPlaceReviews(placeId)
@@ -12,7 +12,7 @@ export default function ListPlaceReview({ placeId }: { placeId: string }) {
 
   return (
     <>
-      <HeaderWithBackButton title='전체 리뷰' />
+      <ActionHeader title='전체 리뷰' isBack />
       <div className='flex flex-col'>
         {reviewData.map((review) => (
           <div key={review.id}>
