+++ src/widgets/header/index.ts
@@ -1,0 +1,3 @@
+export * from './course-plan-header'
+export * from './main-header'
+export * from './action-header'
