+++ src/widgets/index.ts
@@ -1,1 +1,2 @@
 export * from './user-course-list'
+export * from './header'
